Added hardware lock management in Lua RTOS
Now, when a driver is initiated, driver locks the hardware resources that it needs. For example, when LoRa WAN driver is initiated it will lock for a SPI, and 4 GPIO, and when SPI is initiated it lock for 4 GPIO more.

If driver can't lock the resources it needs, an error is returned by the driver, and then, in the corresponding Lua module an exception is generated to the user.