Added Power Bus control in Lua RTOS
Power Bus allows to provide power to sensors connected to a Power Bus. This is useful in IOT deployments to ensure that the sensors do not consume energy while the CPU is sleeping.

Power Bus is automatically on when:

* A sensor read / acquire is done
* A Lora Wan transmission is done

Power Bus is automatically off when:

* CPU enters in sleep mode

Power Bus can be enabled in Kconfig under Lua RTOS / General section.