UART driver bugs
If UART interrupt is fired while an intensive access to SPIFFS (SPI FLASH) is done, some coredumps where appear, due to some functions called in the ISR not placed in IRAM.

For the same reason, trying to stop the execution of the running script, via Ctrl-C causes coredumps, due to that stop signal handlers from Lua are not placed in IRAM. In this case a lot of code is executed in Lua for stop the script, and is not possible to allocate all the needed code in IRAM. For avoid this, Ctrl-C is performed defering the process of Ctrl-C, in the next switch context. For coherence Ctrl-D is performed in the same way.