added captive dns service
captivedns can be started using net.service.captivedns.start()

it will answer all dns queries with a TTL of zero and the address of the
wifi interface.

httpserv has been enhanced to detect foreign host names and redirect to
a pre-defined hostname "config-esp32-settings" in that case - by which
it intentionally also removes the /path and ?data= part of the request.

after connecting to the wifi the android captive login pops up
automatically and can directly be used to configure the esp32's
settings.