'dmesg' shell command causes Guru Meditation Error
dmesg command show the /log/messages.log file contents, but this is only allowed when the filesystem's underlying storage is not the SPI flash. This constraint is due to that all the syslog messages are copied to the /log/messages.log, and if the underlying storage is the SPI flash we can create a hole on the flash, due to SPI flash write-cycle limitations.

The Guru Meditation Error was caused because Lua RTOS try to open (via the fopen function) a NULL filename. Now Lua RTOS makes a sanity check on the filename, and if null or empty, errno is set to NOENT and -1 is returned in the related syscalls:

rename
stat
unlink
mkdir
opendir
readdir
chdir
open