Graphic display driver, and improvements / bugfixes needed
* Added gdisplay driver / gdisplay Lua module

This driver gives the programmer an api to work with graphic displays in a unified form, regardless of the used display.

Main features:

- Optional frame buffer: programmer can enable a frame buffer for render complex scenes and reduce tearing.

- Primitives: pixel, line, rect, filled rect, round rect, filled round rect, circle, filled circle, ellipse, filled ellipse, arc, filled arc, polygon, filled polygon, start, filled star, triangle, filled triangle.

- Fonts: fixed fonts, proportional fonts, custom fonts (loaded from file system).

- Images: bmp (1, 8, 24 bit), jpg.

Supported chipsets:

- ST7735: 1.8", 1.44", 0.96"
- NOKIA 5110

This driver takes lots of source code lines from the work done by Boris Lovošević in his LUA RTOS TFT driver (https://github.com/loboris/Lua-RTOS-ESP32-lobo/tree/master/components/lua_rtos/Lua/modules/screen).

Future work: add ILI9341 / touchpad

* improvement/shell

The shell scan every line typed by user in the console and try to match the line with a shell command. If the line doesn't match with any valid command, the shell searches in the file system for determine if the user has typed a filename, and if so, it will execute the file.

The previous behavior has a side effect when user copy & paste a script to the console. In this case every pasted line doesn't match with any command, and the shell will test if the line corresponds to a filename. This slows down the script execution, because for each pasted line there is a file system access.

This behavior has been removed, and now, if users wants to execute a lua script they must type the command "do filename".

* bugfix/spi driver

Lua RTOS / esp-idf limits the length of an spi transaction when DMA is used to 4 KB. spi_master_op in spi driver has been modified for accept transmissions larger than 4 Kb.

* bugfix/spi driver

esp-idf driver uses DMA, but DMA can't transfer data that resides on FLASH. In this case we copy FLASH data to RAM.

* spi driver improvements

Added a new flag when calling spi_setup for disable dma transfers (SPI_FLAG_NO_DMA). DMA devices and non DMA devices can coexist in the same bus.

* PIN MAP for gdisplay changed to RST GPIO26, CS GPIO15, CMD GPIO2, SPI3. PIN map can be changed in Kconfig. Changes affected to all supported hardware boards: Whitecat N1, Core Board, ESP 32 thing, ang generic.