Added pca9xxx GPIO extender support
* Support for PCA9698, 40 IO, no pull-up
* Support for PCA9505, 40 IO, pull-up
* GPIO driver modified to accept external GPIO using the same API
* gpio_debouncing modified to accept external GPIO using the same API
* ISR registry modified to accept external GPIO using the same API

External GPIO are useful for applications that need a lot number of input / output, such as keyboards, relays, switches, etc.

Please, keep in mind, that when using an external GPIO the GPIO response times are limited by the communication speed between the CPU and the external GPIO. If your application needs to do high speed IO transfers maybe you can't use externa GPIO.

This is an experimental feature. If you want to use this feature you must to enable it in Kconfig, under Lua RTOS -> Hardware -> External GPIO.

Access to external GPIO is done using the same API. GPIO from 0 to 30 are mapped to the internal GPIO and GPIO from 40 to 80 are mapped to the external GPIO.

Examples:

pio.pin.setdir(pio.OUTPUT, 26) --> internal GPIO
pio.pin.setdir(pio.OUTPUT, 50) --> external GPIO

Future work:

It is required to check all supported sensors in Lua RTOS, because not all of them can be used using an external GPIO.