bugfix/memory leak in luaL_driver_error function/error is never free
This function must copy relevant information about the error in the stack space for free the error. This is needed because *error is created in the heap and must be destroy, but relevant information is used when calling to the luaL_error interrupts the program flow and instructions after this call are not executed, so any free call after this call is not executed.