Ethernet driver/added new option in KConfig to select EMAC clock source
The MAC reference clock can be obtained from the ESP32 using the APLL, or from an external 50 Mhz oscillator.When using internal clock the reference clock is put out on a GPIO that must be connected to the PHY's clock input pin. A new option has beeen added in KConfig to select the clock source (LUA_RTOS_ETH_EMAC_CLOCK_SOURCE).