Added an option in KConfig to enable/disable Lua RTOS hardware locks feature
If this feature is enabled, Lua RTOS takes care about all the hardware
resources that your application use. This is useful in the developing
and debuging phase, because in case that your application try to use
a hardware resource that is used in other part of your application, an
exeption will be raised.

This feature consumes arround 3 Kb of memory.