esp-idf/vfs/memory optimization
In esp-idf the dimensions of the VFS are computed from the CONFIG_MAX_FD_BITS constant, and the fact that in newlib the file descriptors are 16-bit wide. In the past, Lua RTOS has changed the CONFIG_MAX_FD_BITS constant to reasonable values to save RAM space.

In addition, the RAM space used by the VFS depends on the VFS_MAX_COUNT constant, which is oversized.

This patch computes the number of used VFS, reducing the used RAM space in 8 Kb.