build system/apply patches if previous patches differ from current patches
This commit apply Lua RTOS patches after upgrading Lua RTOS from the master branch, event if the user does not perform a make clean before.

After applying the patches a 256 hash is computed from the current patch files, and it is stored into the esp-idf/lua_rtos file.
Every time the user builds Lua RTOS the hash is computed again and compared to the stored in the esp-idf/lua_rtos file.
If current hash differs from the stored hash the patches are applied again.