Allow to register a lua module, and autoload the module, or preload the module
Now lua modules can be registered into Lua RTOS as autoload (the module is load on the startup), or as preload (the module is load when user calls to the require function). This featura has been added to the lua-cjson and luasocket library to consume memory only if the libraries are used.

This feature is implemented through the MODULE_REGISTER_RAM and MODULE_REGISTER_ROM macros, in which the last argument set the autoload feature (when set to 1), or the preload feature (when set to 0).