bugfix/i2c driver/2 or more consecutive reads from lua results in an invalid read
The I2C protocol says that for multiple reads the slave sends out data until a NACK and stop condition occurs. To meet this requirement from the Lua I2C module, we must generate an ACK condition for each reading, and generate a dummy NACK reading when stopping the current transaction.