bugfix/spi/assertion failed when setup a spi device under special circumstances
If Lua RTOS can't mount a SDCard configured in SPI mode, the SPI bus remains as it was setup, but really the SPI bus is not setup. After, if the programmer tries to setup a new SPI device that shares the SPI bus with the SDCard the device is not configured propertly, and results in an assertion error.