Added JIT bytecode optimizer, first version
As Lua RTOS makes intensive use of read-only tables, a series of optimizations can be performed on the program's bytecode to speed-up the program execution. This can help that programs written for Lua RTOS to have a similar performance than the writtens in C, and takes a special importance when the programmer use the Lua RTOS hardware-access modules.

Implemented optimizations:

  - A reference to a read-only table field is changed to a constant.
  - A binary expression is changed to a constant if the operands are constants.
  - An unary expressoon is changed to a constant if the operant is a constant.

Notes:

  - This feature is experimental, and disabled by default.
  - To enable the JIT enable the option in KConfig into the "Component config -> Lua RTOS -> Lua -> Tuning" category.
  - When enabled, the readonly cache and lua locks when the program enters the Lua core are disabled.

PLEASE REPORT RELATED ISSUES!!!