DHTxx family sensors refactory
The drivers for the DHTxx family sensors (DHT21, DHT22 & DHT23) have been refactory. This refactory consists in using the ESP32 RMT module to read the data provided by the sensor, which don't require to disable the interrupts when reading data. If all RMT channels are used, the drivers use the traditional bit bang implementation.