Lua net module add callback function to detect interface up / down events
The callback function has been added into the Lua net module. The callback is executed when an interface goes from the down state to the up state, and vice versa. This is useful for applications that need to start network services when an interface becomes up and stop it when an interface becomes down.

In Lua RTOS a network interface is up when it is connected and has an IP, and otherwise it is down.

Usage example:

net.callback(function(event)
  print("event, interface: "..event.interface..", type: "..event.type)
end)