new feature/build system/automatic detection of the serial port where board is attached
For each supported board, Lua RTOS provides a default sdkconfig file, in which the ESPTOOLPY_PORT variable is set, with a standard serial port device name, that can be different on each build platform. This caused the user to have to modify the value of this variable to met the requirements of the platform.

A new feature was added to automatically detect the serial port name in which the board is attached, and avoid to modify the ESPTOOLPY_PORT value manually.

For each board defined in the boards/board.json file, two new fields have been added:

* usb_vid_pid: the expected USB VID:PID expected for the board. Multiple USB VID:PID values can be added, each of them sepparated with the | character.

  To know the USB VID:PID of the available serial ports in your computer you can use the following command in the command line:

  $ python -m serial.tools.list_ports -v

* usb_port_exp: some boards, like the ESP-WROVER-KIT exposes more than 1 serial port for a given USB VID:PID. In this cases the regular expression is used to find the right serial port inspecting the device name, and using the one that matches the regular expression.

If the board has not usb_vid_pid defined, or if there are mora than 1 board attached it is used the serial port defined in board sdkconfig file.