new setting/KConfig/LUA_RTOS_EXTERNAL_ADC_CONNECTED_TO_POWER_BUS
If the external ADC device is connected to the Power BUS, check this option to turn on the bus when the external ADC is attached.

Remember:

Some Lua RTOS compatible boards, such as WHITECAT N2 DEVKIT, have a load switch that can turn on and
off power rails through a GPIO. In Lua RTOS this feature is know as Power BUS, on which you can connect
devices that can be turn on and off on-demand, such as sensors and transceivers that are turn off when
they are not needed to save power consumption.