LoRa WAN stack/avoid joining when MCU have been waked up from deep sleep (see https://github.com/whitecatboard/Lua-RTOS-ESP32/issues/248)
When LoRa session data is obtained by OTAA activation it is stored into RTC memory. If MCU is waked up later from deep sleep, the session is restored, and joining is not required.

For security reasons, and in order to maintain the integrity of the session keys, these are not persistent. This means that if the MCU is waked up from the power, reset, or other states, joining is required.