improve editor for external copy/paste
when copying esp32-external code - e.g. from the PC - to the esp32 via
the editor, the editor's automatic indenting can be *really* annoying.

this commit enables the user to disable the automatic indenting by
pressing [Ctrl]+T, then pasting the external copied code into the editor
and afterwards to re-enable the automatic indenting by pressing [Ctrl]+T
again. automatic indenting is always on by default when launching the
editor, exactly as before.

this commit also enables the user to create a new *empty* file which was
not possible before. launch the editor. press [Ctrl]+S to save. press
[Ctrl]+Q to exit. the file is now saved correctly, unlike before.