callback management functions/callback now accept an argument
All modules that use the callback management functions (luaS_callback_create, luaS_callback_call, ...) now accept an argument in the callback function. This argument is received into the last argument when the callback function is called. This is useful, for example, to receive a reference to an external object when the callback is called.

Example use case:
=================

Implement of end stops in a CNC machine, where end stopts are implementing using the PUSH_SWITCH sensor.

sensor = sensor.attach("PUSH_SWITCH", pin)

sensor:callback(
  function(data, pin)
     print("end stop at pin"..pin..", status = "..data.on)
  end,
  pin
)

In this example the callback now accepts a new argument (the second one) which is the pin where the sensor is attached. Then, the argument is passed as the last argument when the callback is invoked.