Change escaping of raw text to use long string
lua has [[ ]] for raw string embedding
This allow us to not track string quotes and output bigger chunk.
The downside is that we are more likely to hit the buffer size limit of 2k, and miss a lot of preallocated buffer use. So we need to track the length of the line and break.
This change also preserve blank line in raw text.