Lua RTOS stepper driver/bugfix
Add RMT output feedback, configuring RMT output as input/output and attach an ISR to RMT output to compute the real steps.

This bugfix resolves the following scenario:

1) A "x units" movement is started.
2) stepper_stop is called before movement ends

In the above scenario stepper_get_position (introduced in https://github.com/whitecatboard/Lua-RTOS-ESP32/commit/4e2ee11d76a43ac18019b953d4a24b435d3f9088) may return an invalid position due to that steps are precomputed into a buffer, and then are sent to the RMT.

In our proposed solution, an ISR is attached to the RMT output and step updates are done inside, given exactly the number of steps between the start and the end of the RMT transmission.